src/main/java/com/replaymod/replaystudio/pathing/change/AddKeyframe.java
/*
 * Copyright (c) 2021
 *
 * This file is part of ReplayStudio.
 *
 * ReplayStudio is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Lesser General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * ReplayStudio is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public License
 * along with ReplayStudio.  If not, see <https://www.gnu.org/licenses/>.
 */
package com.replaymod.replaystudio.pathing.change;

import com.google.common.base.Preconditions;
import com.google.common.base.Predicates;
import com.google.common.collect.Iterables;
import com.replaymod.replaystudio.pathing.path.Keyframe;
import com.replaymod.replaystudio.pathing.path.Path;
import com.replaymod.replaystudio.pathing.path.Timeline;
import lombok.NonNull;

/**
 * Adds a new property.
 */
public final class AddKeyframe implements Change {
    @NonNull
    public static AddKeyframe create(Path path, long time) {
        return new AddKeyframe(path.getTimeline().getPaths().indexOf(path), time);
    }

    AddKeyframe(int path, long time) {
        this.path = path;
        this.time = time;
    }

    /**
     * Path index
     */
    private final int path;

    /**
     * Time at which the property should be injected.
     */
    private final long time;

    /**
     * Index of the newly created property.
     */
    private int index;

    private boolean applied;

    @Override
    public void apply(Timeline timeline) {
        Preconditions.checkState(!applied, "Already applied!");

        Path path = timeline.getPaths().get(this.path);
        Keyframe keyframe = path.insert(time);
        index = Iterables.indexOf(path.getKeyframes(), Predicates.equalTo(keyframe));

        applied = true;
    }

    @Override
    public void undo(Timeline timeline) {
        Preconditions.checkState(applied, "Not yet applied!");

        Path path = timeline.getPaths().get(this.path);
        path.remove(Iterables.get(path.getKeyframes(), index), true);

        applied = false;
    }
}
